+++ api/lib/subscription/slashing.js
@@ -21,8 +21,8 @@ class SlashingMonitor {
   // to prevent adding a slash twice we filter the slashes
   storeSlashes(filterReason) {
     return (tendermintResponse) => {
-      const coinLookup = network.getCoinLookup(
-        network,
+      const coinLookup = this.network.getCoinLookup(
+        this.network,
         this.network.stakingDenom,
         `viewDenom`
       )
@@ -106,7 +106,7 @@ class SlashingMonitor {
             })
           )
           database(config)('').upsert('slashes', rows)
-          console.log('Wrote missed block events', rows.length)
+          // console.log('Wrote missed block events', rows.length)
         } catch (error) {
           console.error('Failed to write missed block events', error)
           Sentry.captureException(error)